add prototype method for stopping delays
Blueprints can use this variable to key behavior off. In this case if player doesnt choose, a designer can define portrayals using delays that do not progress if the bool is set to false.

Enabling a player choice to exit the inaction portrayals execution inside the level Blueprint.

WIP method, currently delay will still run, but a macro can be defined to check our subsystems bool and not continue if set to false.